+++ test/bulk_insert/worker_test.rb
@@ -240,6 +240,26 @@ class BulkInsertWorkerTest < ActiveSupport::TestCase
     assert_equal mysql_worker.compose_insert_query, "INSERT IGNORE INTO \"testings\" (\"greeting\",\"age\",\"happy\",\"created_at\",\"updated_at\",\"color\") VALUES ('Yo',15,'f',NULL,NULL,'chartreuse')"
   end
 
+  test "adapter dependent mysql methods work for mysql2" do
+    mysql_worker = BulkInsert::Worker.new(
+      Testing.connection,
+      Testing.table_name,
+      %w(greeting age happy created_at updated_at color),
+      500, # batch size
+      true, # ignore
+      true) # update_duplicates
+    mysql_worker.adapter_name = 'Mysql2'
+
+    assert_equal mysql_worker.adapter_name, 'Mysql2'
+    assert_equal (mysql_worker.adapter_name == 'Mysql2'), true
+    assert_equal mysql_worker.ignore, true
+    assert_equal ((mysql_worker.adapter_name == 'Mysql2') & mysql_worker.ignore), true
+
+    mysql_worker.add ["Yo", 15, false, nil, nil]
+
+    assert_equal mysql_worker.compose_insert_query, "INSERT IGNORE INTO \"testings\" (\"greeting\",\"age\",\"happy\",\"created_at\",\"updated_at\",\"color\") VALUES ('Yo',15,'f',NULL,NULL,'chartreuse') ON DUPLICATE KEY UPDATE greeting=VALUES(greeting), age=VALUES(age), happy=VALUES(happy), created_at=VALUES(created_at), updated_at=VALUES(updated_at), color=VALUES(color)"
+  end
+
   test "adapter dependent postgresql methods" do
     pgsql_worker = BulkInsert::Worker.new(
       Testing.connection,
@@ -277,5 +297,19 @@ class BulkInsertWorkerTest < ActiveSupport::TestCase
     sqlite_worker.add ["Yo", 15, false, nil, nil]
 
     assert_equal sqlite_worker.compose_insert_query, "INSERT OR IGNORE INTO \"testings\" (\"greeting\",\"age\",\"happy\",\"created_at\",\"updated_at\",\"color\") VALUES ('Yo',15,'f',NULL,NULL,'chartreuse')"
+  end
+
+  test "mysql adapter can update duplicates" do
+    mysql_worker = BulkInsert::Worker.new(
+      Testing.connection,
+      Testing.table_name,
+      %w(greeting age happy created_at updated_at color),
+      500, # batch size
+      false, # ignore
+      true) # update_duplicates
+    mysql_worker.adapter_name = 'MySQL'
+    mysql_worker.add ["Yo", 15, false, nil, nil]
+
+    assert_equal mysql_worker.compose_insert_query, "INSERT  INTO \"testings\" (\"greeting\",\"age\",\"happy\",\"created_at\",\"updated_at\",\"color\") VALUES ('Yo',15,'f',NULL,NULL,'chartreuse') ON DUPLICATE KEY UPDATE greeting=VALUES(greeting), age=VALUES(age), happy=VALUES(happy), created_at=VALUES(created_at), updated_at=VALUES(updated_at), color=VALUES(color)"
   end
 end